Changes packages in part2-parts from simple to test.featurestart.
