+++ dev/build.featureStart.part2_fat/fat/src/com/ibm/ws/test/featurestart/FeaturesStartTest2.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2019, 2022 IBM Corporation and others.
+ * Copyright (c) 2019, 2023 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License 2.0
  * which accompanies this distribution, and is available at
@@ -10,13 +10,13 @@
  * Contributors:
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
-package com.ibm.ws.simple;
+package com.ibm.ws.test.featurestart;
 
 import org.junit.BeforeClass;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
-import com.ibm.ws.simple.base.FeaturesStartTestBase;
+import com.ibm.ws.test.featurestart.FeaturesStartTestBase;
 
 import componenttest.annotation.Server;
 import componenttest.custom.junit.runner.FATRunner;
@@ -35,26 +35,26 @@ import componenttest.topology.impl.LibertyServer;
  * extra running room before a new split is necessary.
  */
 @RunWith(FATRunner.class)
-public class FeaturesStartTest3 extends FeaturesStartTestBase {
-    public static final String SERVER_NAME_3 = "features.start.3.server";
+public class FeaturesStartTest2 extends FeaturesStartTestBase {
+    public static final String SERVER_NAME_2 = "features.start.2.server";
 
-    @Server(SERVER_NAME_3)
-    public static LibertyServer server3;
+    @Server(SERVER_NAME_2)
+    public static LibertyServer server2;
 
     public static final int NUM_BUCKETS = 4;
-    public static final int BUCKET_NO = 3;
-    public static final int SPARSITY = 1;
+    public static final int BUCKET_NO = 2;
+    public static final int SPARSITY = 0;
 
     @BeforeClass
     public static void setUp() throws Exception {
-        FeaturesStartTestBase.setParameters(FeaturesStartTest3.class,
-                                            server3, SERVER_NAME_3,
+        FeaturesStartTestBase.setParameters(FeaturesStartTest2.class,
+                                            server2, SERVER_NAME_2,
                                             NUM_BUCKETS, BUCKET_NO, SPARSITY);
         FeaturesStartTestBase.setUp();
     }
 
     @Test
-    public void testStartFeaturesPart3() throws Exception {
+    public void testStartFeaturesPart2() throws Exception {
         FeaturesStartTestBase.testStartFeatures();
     }
 }
