+++ dev/build.featureStart.part4_fat/fat/src/com/ibm/ws/test/featurestart/FATSuite.java
@@ -10,14 +10,14 @@
  * Contributors:
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
-package com.ibm.ws.simple;
+package com.ibm.ws.test.featurestart;
 
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
 import org.junit.runners.Suite.SuiteClasses;
 
 @RunWith(Suite.class)
-@SuiteClasses({ FeaturesStartTest3.class })
+@SuiteClasses({ FeaturesStartTest4.class })
 public class FATSuite {
     // EMPTY
 }
